+++ fabric/src/main/java/io/github/sakurawald/module/initializer/tpa/config/model/TpaConfigModel.java
@@ -1,8 +1,12 @@
 package io.github.sakurawald.module.initializer.tpa.config.model;
 
+import com.google.gson.annotations.SerializedName;
 import io.github.sakurawald.core.job.impl.MentionPlayersJob;
 
 public class TpaConfigModel {
-    public int timeout = 300;
+
+    @SerializedName(value = "request_timeout", alternate = "timeout")
+    public int request_timeout = 300;
+
     public MentionPlayersJob.MentionPlayer mention_player = new MentionPlayersJob.MentionPlayer();
 }