+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
-- (< MC 1.21.5) [deathlog] fix: should not restore the armor slots in reversed order.
-- [deathlog] feature: add death node information as lore of `restore` button.
+- [tpa] fix: should send a feedback message to the `receiver` if the request is `accepted`.
+- [tpa] i18n: new lang keys for `accept` and `reject` buttons.